+++ projects/angular-django2/src/lib/angular-django2.ts
@@ -1,4 +1,0 @@
-export * from './angular-django2-config';
-export * from './angular-django2.providers';
-export * from './angular-django2.service';
-export * from './angular-django2.tokens';