+++ projects/angular-django2/src/public-api.ts
@@ -1,5 +1,0 @@
-/*
- * Public API Surface of angular-django2
- */
-
-export * from './lib/angular-django2';